+++ changelog.md
@@ -98,8 +98,8 @@
 ### 5.4.0
 * support `.paragraphs()` - big but non-breaking change.
 * :warning: changes to output of `.json()`. cleaning-up redundant data. (sorry for the frequent changes here).:warning:
-* * remove top-level `templates` data (found in `section`)
-* * remove top-level `citations` data (found in `section`)
+* * remove top-level `templates` data (found in `section`) - resume it with `{templates:true}`
+* * remove top-level `citations` data (found in `section`) - resume it with `{citations:true}`
 * return empty arrays in `.json()` again  ¯\_(:/)_ /¯
 * remove <h1> title on html output
 * change ambiguous `options.title` to `options.headers`